// src/services/buyer/messages/typing.d.ts
export interface Message {
  name: string;
  time: string;
  snippet: string;
  avatar: string;
}

export type MessageSender = 'me' | 'them';

export interface ChatMessage {
  id: number;
  text?: string;
  image?: string;
  sender: MessageSender;
  time: string;
}

export interface Conversation {
  id: number;
  name: string;
  avatar: string;
  lastMsg: string;
  time: string;
  online: boolean;
  lastActive?: string;
  messages: ChatMessage[];
}